新增清除redis的定时任务

+++ web-api/src/main/java/com/example/workpush/controller/JobController.java
@@ -1,12 +1,10 @@
 package com.example.workpush.controller;
 
 import cn.hutool.core.lang.Validator;
-import com.example.workpush.quartzJob.CleanRedisIDJob;
 import com.example.workpush.quartzJob.WorkPushJob;
 import jakarta.annotation.Resource;
 import lombok.extern.slf4j.Slf4j;
 import org.quartz.*;
-import org.quartz.impl.StdSchedulerFactory;
 import org.quartz.impl.matchers.GroupMatcher;
 import org.springframework.web.bind.annotation.*;
 
@@ -35,7 +33,12 @@ public class JobController {
             return "岗位类型输入错误";
         }
         try {
-
+            JobKey jobKey = new JobKey("workPushJob" + to + categoryType + key, "jobGroup");
+            // 判断当前jobKey是否存在
+            boolean isExist = scheduler.checkExists(jobKey);
+            if(isExist) {
+                return "当前定时任务已存在，无需重复设置";
+            }
             // 开启工作推送任务
             JobDataMap jobDataMap = new JobDataMap();
             jobDataMap.put("to", to);
@@ -43,35 +46,23 @@ public class JobController {
             jobDataMap.put("key", key);
             JobDetail jobDetail1 = JobBuilder.newJob(WorkPushJob.class)
                     .withDescription("开启工作推送服务")
-                    .withIdentity("workPushJob" + to + key,"jobGroup")
+                    .withIdentity(jobKey)
                     .storeDurably(true)
                     .setJobData(jobDataMap)
                     .build();
-            CronScheduleBuilder cron1 = CronScheduleBuilder.cronSchedule("0/10 * * * * ?");
-            Trigger trigger1 = TriggerBuilder.newTrigger()
+            CronScheduleBuilder cron = CronScheduleBuilder.cronSchedule("0/10 * * * * ?");
+            TriggerKey triggerKey = new TriggerKey("workPushTrigger" + to + categoryType + key, "jobGroup");
+            Trigger trigger = TriggerBuilder.newTrigger()
                     .withDescription("工作推送触发器")
-                    .withIdentity("workPushTrigger" + to + key,"defaultGroup")
-                    .withSchedule(cron1)
+                    .withIdentity(triggerKey)
+                    .withSchedule(cron)
                     .build();
-            // 清除redis中的数据
-//            JobDetail jobDetail2 = JobBuilder.newJob(CleanRedisIDJob.class)
-//                    .withDescription("清除redis")
-//                    .usingJobData("to", to)
-//                    .build();
-//            CronScheduleBuilder cron2 = CronScheduleBuilder.cronSchedule("0 0 0 * * ?");
-//            Trigger trigger2 = TriggerBuilder.newTrigger()
-//                    .withDescription("清除redis触发器")
-//                    .withIdentity("cleanRedisTrigger","defaultGroup")
-//                    .withSchedule(cron2)
-//                    .build();
-
-            scheduler.scheduleJob(jobDetail1, Set.of(trigger1), true);
-//            scheduler.scheduleJob(jobDetail2, Set.of(trigger2), true);
+            scheduler.scheduleJob(jobDetail1, Set.of(trigger), true);
 
             scheduler.start();
 
             return "正确开启推送！";
-        } catch (SchedulerException e) {
+        } catch (Exception e) {
             throw new RuntimeException(e);
         }
     }
@@ -107,7 +98,12 @@ public class JobController {
             return "岗位类型输入错误";
         }
         try {
-            JobKey jobKey = new JobKey("workPushJob" + to + key, "jobGroup");
+            JobKey jobKey = new JobKey("workPushJob" + to + categoryType + key, "jobGroup");
+            // 判断当前jobKey是否存在
+            boolean isExist = scheduler.checkExists(jobKey);
+            if(!isExist) {
+                return "当前定时任务不存在，无需取消";
+            }
             scheduler.deleteJob(jobKey);
         } catch (SchedulerException e) {
             throw new RuntimeException(e);
